Backend/ml_service/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
xgboost
joblib
numpy
scipy
supabase
python-dotenv
google-generativeai
